Add adjustable text size feature
A new adjustable text size feature has been added to the sidebar of the portfolio project. This feature will let the user increase or decrease the size of the text on the site for a more personalized reading experience. Two new buttons have been added, '-' and '+', which decrease and increase the text size respectively. Also, the portfolio's CSS file has been updated to accommodate this feature. Furthermore, a few changes in the existing code were made to enhance code readability.