Refactor variable scope and async function calls
The changes made brought the IPSTACK_ACCESS_KEY constant into the getIpStack function where it is being used. Every occurrence of promises (ipStackPromise, worldTimePromise, sunPromise) directly used in the code were replaced with the async functions they were assigned to. This is done to ensure encapsulation of the IPSTACK_ACCESS_KEY constant and to simplify the structure, utilizing async/await more directly.